backend/requirements.txt
openai>=1.14.0
fastapi>=0.111.0
uvicorn>=0.29.0
requests>=2.31.0
pydantic[email]>=2.7.0
python-dotenv>=1.0.0
redis>=5.0.0
urllib3>=2.2.0
